+++ src/main/java/com/furnitureapp/repository/clientele/CustomerRepository.java
@@ -3,5 +3,5 @@ package com.furnitureapp.repository.clientele;
 import com.furnitureapp.entity.clientele.Customer;
 import com.furnitureapp.repository.IRepository;
 
-public interface CustomerRepository extends IRepository<Customer, Long> {
+public interface CustomerRepository extends IRepository<Customer, Integer> {
 }
